===== v0.5.1
1. API changes to match SNIA standard
2. API sync mode for existing operation types
3. Sample testing code updated
4. Error handling in API
5. Bug fix

+++ PDK/core/tools/install_deps.sh
@@ -1,11 +1,28 @@
+#!/bin/bash
+# Please run this script as root.
 
-sudo apt-get install libcunit1-dev
-sudo apt-get install libaio-dev
-sudo apt-get install libtbb-dev
-sudo apt-get install openssl
+if [ -s /etc/redhat-release ]; then
+    # Includes Fedora, CentOS
+    sudo yum install CUnit-devel
+    sudo yum install libaio-devel
+    sudo yum install tbb
+    sudo yum install openssl
 
-#boost-program-options
+    sudo yum install libev-devel
 
+elif [ -f /etc/debian_version ]; then
+    # Includes Ubuntu, Debian
+    sudo apt-get install libcunit1-dev
+    sudo apt-get install libaio-dev
+    sudo apt-get install libtbb-dev
+    sudo apt-get install openssl
 
-#aerospike
-sudo apt-get install libev-dev
+    #boost-program-options
+
+    #aerospike
+    sudo apt-get install libev-dev
+
+else
+    echo "unsupported system type."
+    exit 1
+fi
